Performs dplyr if_else

This function is similar to R function rename.

caseWhenDS(dynamic_dots = NULL, .default = NULL, .ptype = NULL, .size = NULL)

Arguments

dynamic_dots

A sequence of two-sided formulas. The left hand side (LHS) determines which values match this case. The right hand side (RHS) provides the replacement value. The LHS inputs must evaluate to logical vectors. The RHS inputs will be coerced to their common type. All inputs will be recycled to their common size. That said, we encourage all LHS inputs to be the same size. Recycling is mainly useful for RHS inputs, where you might supply a size 1 input that will be recycled to the size of the LHS inputs. NULL inputs are ignored.

.default

The value used when all of the LHS inputs return either FALSE or NA.

.ptype

An optional prototype declaring the desired output type. If supplied, this overrides the common type of true, false, and missing.

.size

An optional size declaring the desired output size. If supplied, this overrides the size of condition.

Value

the object specified by the newobj argument of ds.case_whenwhich is written to the serverside.

Details

Performs dplyr rename